Support false value for cell grow attribute
