+++ lml/src/main/java/com/github/czyzby/lml/parser/impl/attribute/table/cell/CellExpandXLmlAttribute.java
@@ -15,15 +15,4 @@ public class CellExpandXLmlAttribute extends AbstractCellLmlAttribute {
             final String rawAttributeData) {
         cell.expand(parser.parseBoolean(rawAttributeData, actor), determineExpandY(cell));
     }
-
-    protected boolean determineExpandY(final Cell<?> cell) {
-        try {
-            return cell.getExpandY() > 0;
-        } catch (final Exception exception) {
-            // LibGDX Scene2D method returns int, while the field is an Integer that might not have been initiated. This
-            // causes a NPE - so when an exception is thrown, we assume that the expand was not set.
-            Exceptions.ignore(exception);
-            return false;
-        }
-    }
 }
