+++ lml/src/main/java/com/github/czyzby/lml/parser/impl/attribute/table/cell/CellFillXLmlAttribute.java
@@ -15,15 +15,4 @@ public class CellFillXLmlAttribute extends AbstractCellLmlAttribute {
             final String rawAttributeData) {
         cell.fill(parser.parseBoolean(rawAttributeData, actor), determineFillY(cell));
     }
-
-    protected boolean determineFillY(final Cell<?> cell) {
-        try {
-            return cell.getFillY() > 0f;
-        } catch (final Exception exception) {
-            // LibGDX Scene2D method returns float, while the field is a Float that might not have been initiated. This
-            // causes a NPE - so when an exception is thrown, we assume that the fill was not set.
-            Exceptions.ignore(exception);
-            return false;
-        }
-    }
 }
