+++ lml/src/main/java/com/github/czyzby/lml/parser/impl/attribute/table/cell/CellGrowLmlAttribute.java
@@ -14,6 +14,9 @@ public class CellGrowLmlAttribute extends AbstractCellLmlAttribute {
             final String rawAttributeData) {
         if (parser.parseBoolean(rawAttributeData, actor)) {
             cell.grow();
+        } else {
+            cell.expand(false, false);
+            cell.fill(false, false);
         }
     }
 }
